Project_source_files
The goal of this project was to evaluate various approaches for using
text features to do data analytics.  The data set that was used is from
the 2014 KDD cup Kaggle dataset. The contest focused on data obtained
from DonorsChoose.org, a site dedicated to providing a platform for
people to donate towards classroom projects.  The target variable for
the contest was if a particular campaign was “exciting” and therefore
should have more resources and focus put on it.  There were several
characteristics that contributed to the label of “exciting” such as if a
campaign was eventually fully funded.